requirements.txt
rich
python-github
python-dotenv
